Modernized to current version of SvelteKit
