+++ .gitignore
@@ -4,3 +4,4 @@
 /docs/
 /types/
 /pnpm-lock.yaml
+.idea/
